portal_with_advertisements/requirements.txt
Django==3.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
factory-boy==3.2.0
Markdown==3.3.4
